Changed float() to int()

+++ simple_projects_in_python_2/program_for_dividing_two_numbers_without_decimal_point.py
@@ -7,7 +7,7 @@ num_2 = int(input("Input your second number: "))
 # check if the second number is not equal to zero
 while num_2 == 0:
     print("Cannot divide by zero.")
-    num_2 = float(input("Input your second number: "))
+    num_2 = int(input("Input your second number: "))
 
 # num_1 / num_2
 # print the quotient without decimal point
